Move code editor and button to separate component, fix pushNewDefinition call
