+++ app/src/editor/canvas/components/CodeEditor/CodeEditor.pcss
@@ -1,0 +1,15 @@
+.button {
+  border: 1px solid #EFEFEF;
+  border-radius: 2px;
+  background: white;
+  color: #525252;
+  font-size: 10px;
+  font-weight: 500;
+  letter-spacing: 0;
+  line-height: 16px;
+  width: 100%;
+}
+
+.button:focus {
+  outline: none;
+}
