+++ app/src/editor/canvas/components/CodeEditor/DebugWindow.pcss
