+++ app/src/editor/shared/components/modules/Custom.pcss
@@ -2,19 +2,3 @@
   line-height: 16px;
   padding: 8px !important;
 }
-
-.button {
-  border: 1px solid #EFEFEF;
-  border-radius: 2px;
-  background: white;
-  color: #525252;
-  font-size: 10px;
-  font-weight: 500;
-  letter-spacing: 0;
-  line-height: 16px;
-  width: 100%;
-}
-
-.button:focus {
-  outline: none;
-}
